pyproject.toml
[tool.poetry]
name = "recsys2020-challenge"
version = "0.1.0"
description = ""
authors = ["Naomichi Agata <agatan039@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.3"
pandas = "^1.0.1"
pandas-gbq = "^0.13.1"
google-cloud-storage = "^1.26.0"
scikit-learn = "^0.22.2"
numpy = "^1.18.1"
scipy = "^1.4.1"
matplotlib = "^3.2.0"
click = "^7.1.1"
torch = "^1.5.0"
transformers = "^2.9.0"
apache-beam = {extras = ["gcp"], version = "^2.19.0"}
pyarrow = "0.15.1"
google-cloud-bigquery-storage = {version = "^0.8.0", allow-prereleases = true}
lightgbm = "^2.3.1"
tensorflow = "^2.2.0"
networkx = "^2.4"

[tool.poetry.dev-dependencies]
ipython = "^7.13.0"
pylint = "^2.4.4"
black = "^19.10b0"
mypy = "^0.761"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
